refactor(tui): move archived context rendering
Extract archived-context parsing and rendering from the main history renderer into history/archived_context.rs while keeping history_cells_from_message and rendering behavior unchanged.

Refs #3308.
